Optimize ChaCha20-Poly1305 initialization
The ChaCha20 initialization cost has been
found to be expensive in the OpenSSL 3.x
API compared to the OpenSSL 1.x API.

In order to optimize the cost on OpenSSL
3.x, a flag is setup to distinguish two
different types of initializations. One
initialization would be a full
initialization. It will perform the full
initialization when the mode is switched
between encrypt and decrypt, or a first
time encryption or decryption. The other
initialization would be a partial
initialization. Once the mode is decided,
the partial initialization will only use
key and iv, but will NOT recreate and
reinitialize the EVP context as what we
have for now. The key and iv are required
once per Cipher instance, the EVP context
can be used whenever we are reusing a
specific Java cipher object within methods
such as Cipher.doFinal().

Signed-off-by: Jinhang Zhang <Jinhang.Zhang@ibm.com>