Get CKA_VALUE from derived secretkey in FIPS mode
Currently we can not get the CKA_VALUE
from a derived secretkey in FIPS mode
during the keyagreement. It happens
because the native C_GetAttributeValue
function treat the CKA_VALUE attribute
as a sensitive one.

Our solution is using the Java code to
replace the native C_GetAttributeValue
function to get the CKA_VALUE from the
derived secretkey.